requirements.txt
akshare
pandas
matplotlib
mplfinance
pandas_ta
streamlit
mysql-connector-python
bcrypt
